@@ -21,9 +21,10 @@ import {Component, HostListener, OnDestroy, OnInit} from '@angular/core';
 import {Subscription} from 'rxjs';
 
 @Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss'],
+    selector: 'app-root',
+    templateUrl: './app.component.html',
+    styleUrls: ['./app.component.scss'],
+    standalone: false
 })
 export class AppComponent implements OnInit, OnDestroy {
   isLive = false;
@@ -34,7 +35,7 @@ export class AppComponent implements OnInit, OnDestroy {
   }
 
   @HostListener('document:visibilitychange', ['$event'])
-  onPageVisible(): void {
+  onPageVisible(event: Event): void {
     if (document.hidden) {
       this.apiService.isWindowOpenedSubject.next(false);
     } else {
