@@ -23,9 +23,10 @@ import {SearchService} from '../oer-server/search/search.service';
 import {Router} from '@angular/router';
 
 @Component({
-  selector: 'app-nav',
-  templateUrl: './nav.component.html',
-  styleUrls: ['./nav.component.scss']
+    selector: 'app-nav',
+    templateUrl: './nav.component.html',
+    styleUrls: ['./nav.component.scss'],
+    standalone: false
 })
 export class NavComponent implements OnInit, OnDestroy {
   currentSearchPhrase = '';
