@@ -19,8 +19,9 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-not-found',
-  templateUrl: './not-found.component.html',
+    selector: 'app-not-found',
+    templateUrl: './not-found.component.html',
+    standalone: false
 })
 export class NotFoundComponent {
 
