@@ -19,7 +19,8 @@
 import {Pipe, PipeTransform} from '@angular/core';
 
 @Pipe({
-  name: 'search'
+    name: 'search',
+    standalone: false
 })
 export class SearchPipe implements PipeTransform {
 
