@@ -26,9 +26,10 @@ import {Observable, Subscription} from 'rxjs';
 import {first} from 'rxjs/operators';
 
 @Component({
-  selector: 'app-search',
-  templateUrl: './search.component.html',
-  styleUrls: ['./search.component.scss']
+    selector: 'app-search',
+    templateUrl: './search.component.html',
+    styleUrls: ['./search.component.scss'],
+    standalone: false
 })
 export class SearchComponent extends AbstractReadMoreComponent implements OnInit, OnDestroy {
   programEntryList: ProgramEntry[] = [];
