@@ -18,19 +18,31 @@
  */
 
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { ApiService } from '../api.service';
+import { BehaviorSubject } from 'rxjs';
 
 import { XmltvComponent } from './xmltv.component';
 
 describe('XmltvComponent', () => {
   let component: XmltvComponent;
   let fixture: ComponentFixture<XmltvComponent>;
+  let apiServiceMock: any;
 
   beforeEach(async () => {
+    apiServiceMock = {
+      updateStatus: jasmine.createSpy('updateStatus'),
+      statusSubject: new BehaviorSubject<any>(null),
+      isWindowOpenedSubject: new BehaviorSubject<boolean>(true)
+    };
+
     await TestBed.configureTestingModule({
-      imports: [XmltvComponent]
+      declarations: [XmltvComponent],
+      providers: [
+        { provide: ApiService, useValue: apiServiceMock }
+      ]
     })
     .compileComponents();
-    
+
     fixture = TestBed.createComponent(XmltvComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
